huge amount of restructuring. DOM-manipulation.js module added, tidied all js code.
